split flow.pyx in 3
