+++ src/itzi/compute/rastermetrics.pyx
